Use -funwind-tables when building target libraries
This patch adds the -funwind-tables flag when building our prebuilt
libraries in order to ensure that the generated code can easily be
unwound at runtime.

This prevents stack-traces from being incorrectly shortened, for
example when using the malloc leak checked, the STLport/GABI++
operator new implementation could not be unwound, and the stack
trace could not show where the code was really allocated/freed
in the original source code.

Change-Id: I424754c4d891416b34ba1f67099fbed62252286e