Fixed LOCAL_SHORT_COMMANDS to not recompile/relink source/object every time
When LOCAL_SHORT_COMMANDS is set to true, compiler options are emitted to a
response file and read off using the '@file' syntax.  This is to support very
long commmand line w/o running over the limit of shell command buffer.

Obviously target depends on the response file which shall be regenerated
every time to account for potential changes in command line options.
Unfortunately this triggers rebuilding of target as well even when the content
of response file is identical to the last one.

This CL fixes it by emitting options to a temp file, and only update the response
file when contents differ

Change-Id: Ia2be6df7bbc0d0816b60d552f3fde004f42ca748