Make LOCAL_SHORT_COMMANDS work on Darwin.
Our implementation of LOCAL_SHORT_COMMANDS relied on 'echo -n' to append
data to an existing file, but this option is not supported by Darwin's
echo program.

Modify our build scripts to use "printf %s" instead to do the same thing
on Posix systems. On Windows, which doesn't have printf, keep using the
toolbox echo program that supports the flag.

Change-Id: I35e14279a60e2dd54a0078cdcc5e177984e155d8