Script build-gcc.sh builds gold as the default linker
Modify build-gcc.sh to build gold for X86/ARM on Linux, MaxOSX,
and Windows hosts, and set it as the default linker.

Bump up TOOLCHAIN_GIT_DATE to get fix in toolchain which allows
ld.gold to be built in Windows.

Put i586-mingw32msvc- in the last of mingw toolchain list,
because it may crash during ld.gold compilation

Change-Id: I779b734711693c8b4259919da9b4a442fe9c7cdd