Backport ".init_array" patches from the upstream.
Android dynamic linker does not support .ctors/.dtors,
thus we have to backport .init_array/.fini_array
from the upstream.

Besides, this CL fixes the regression of test_gnustl_locale
when we are using clang.

Some portion of this CL is basing on Rafael Espindola's
works, which were submitted to LLVM upstream:

  commit d6b43a317e71246380db55a50b799b062b53cdce
  Author: Rafael Espindola <rafael.espindola@gmail.com>
  Date:   Tue Jun 19 00:48:28 2012 +0000

  commit 8af669f2f1d92436fe6dc43144bb084a620e7516
  Author: Rafael Espindola <rafael.espindola@gmail.com>
  Date:   Tue Jun 19 01:26:10 2012 +0000

Change-Id: I0c202ef0d095d1dcfe685d3bff6c3cac55da4929