Add -target to clang in NDK standalone mode
To make clang/clang++ easier drop-in replacements for gcc/g++,
add predefined -target for clang/++ in NDK standalone mode.

  clang        # a script
  clang++      # a script
  clang3.1     # copied from toolchains/llvm-3.1/prebuilt/<host>/bin/clang
  clang3.1++ -> clang3.1    # linked to clang3.1

where clang contains the following line (assuming arm):

  `dirname $0`/clang3.1 -target armv5te-none-linux-androideabi "$@"

Likewise for clang++ in arm

  `dirname $0`/clang3.1++ -target armv5te-none-linux-androideabi "$@"

Note that for arm, clang will change target based on the presence of
subsequent -march=armv7-a and/or -mthumb

eg.
  With -march=armv7-a, -target becomes armv7-none-linux-androideabi
  With -mthumb, -target becomes thumb-none-linux-androideabi
  With both, -target becomes thumbv7-none-linux-androideabi

Change-Id: Id5dc08091ecf9727d383c4fa4c19a812ca0449da