Remove undefined behavior in exceptions-crash test.
Not returing a value in a function with non-void
return type will result in undefined behavior.

NOTE: This issue is discovered when I was testing the
new version of Clang.  It will add an UND instruction
which will raise SIGILL in the debug build to the end
of function with non-void return type and without a
return statement.

Change-Id: Ib5bdc00478e19ee02fb84aec7a1925f89120c37a