Fix a bug preventing STL dependencies to certain C++ sources.
ndk-build was modified recently to recognize C++ source files
that don't use the .cpp extension. However, the module-get-cpp-sources
definition in build/core/definitions.mk was not updated properly.

The end result is that if a module only used C++ source files that
was not .cpp, it would get built with the C++ compiler (correctly),
but the APP_STL setting would not be applied to it, result in a build
error (for example, when including a C++ STL header).

This patch fixes this, and also adds a regression test for this issue.

Change-Id: Ib68d07492bdbb597e5c6f414cc989eb816ad8752