Set LLVM bug report url to Android issue tracker.
Since we are distributing the modified LLVM/Clang in
the NDK, it will be better not to suggest the end-user
to report bug to LLVM upstream.  It is preferred to
suggest the user to report to our issue tracker, so that
we can test whether it is our bug or an upstream bug.

Change-Id: I7e1ab3b0dd203c1fbe485e4bde560689a85f720c