cpu-features: Add android_setCpu()
This patch adds a new function to the cpu-features library to be able
to set the CPU properties manually, instead of relying on parsing /proc.

This is mostly useful on ARM for sandboxed processes in Android 4.1
and higher, because these don't have permissions to access /proc.

For more context, see http://code.google.com/p/chromium/issues/detail?id=164154

Change-Id: I1c4f22360210bcc3276093d2c81fa8d6c1e0f146