Added ndk-gdb{.py|-py|-py.cmd}
Pythonic replacements for the ndk-gdb shell script.

Attempts to emulate the old behaviour, including verbose mode and
logging and also the general layout of the shell script.

Added a new command line option, --tui/-t to launch
gdb in text user interface mode, if the gdb is built
with support for this.

Change-Id: Ic02feb18d70fe449bf54a531ffb973513dc82783