Modify gabi++ to conform better to libcxxabi API.
This change focuses on making gabi++ source compatible with libcxx.
It introduces a new macro, GXXABI_LIBCXX which conditionally enables
and disables functions in gabi++ to make the exported API behave like
libcxxabi's. It also refactors gabi++ somewhat to remove symbols that
shouldn't really be in namespace std.

After this CL, it is possible to make a running Hello World using
libcxx on Android.

Change-Id: I3736cb6471ecd0685c1528f904c0901fdfd8f441