Push the test executables just before we run it.
We used to pushing many executables to the devices at once.
It may take long time for some test.  As a result, we can't
distinguish the long file transfering time or the device
hang.

In this CL, we slightly changed the run-tests.sh so that
we can push the executables just before the execution.

Change-Id: I972c7181e7cd833108523cc9707807e8eaf86ad2