Unhide EH helper needed by libgcc.a
ndk-build puts another libgcc.a between user's *.o/*.a
and *.so.  Reference to EH helper functions such as
__cxa_type_match, __cxa_begin_cleanup and __cxa_call_unexpected
may be copied to the generated *so if code in *.o/*.a throws
exception.  If later this *so is linked by other executable/
libraies with the same libgcc.a formation, the hidden EH
helpers in libstlport/libgnustl_static.a will be referenced,
which causes linker warning, and may halt compilation in case
of -Wl,--fatal-warnings

See internal bug ref b/8247455

Change-Id: I9c55fab5de0a379b9a3ffd5d9e9d39891ebf83bc