Add LOCAL_THIN_ARCHIVE support.
This adds support for generating "thin archives" when building
static libraries. Instead of containing a copy of their object
files, these contain just pointers/paths to them, making the file
much smaller and avoiding data duplication, reducing the size
of the build output.

The drawback is that these libraries cannot generally be moved,
or will fail at link time if the object files are removed.

BUG=40303

Change-Id: Iee8fc34d9a2b197dac867bb87f6a6977e3ad1443