PyPrPr :: Added --stdcxx-py-pr argument and support to ndk-gdb.py
This is the last commit needed to enable Python pretty-printers
in ndk-gdb.py

Valid options are 'auto', 'none', 'stlport', 'gnustdcxx' and
'gnustdcxx-$VER' where $VER is a foldername in:
$NDK/prebuilt/$SYSTEM/share/pretty-printers/libstdcxx with 'gcc-'
removed.

'auto' mode works by looking at the .so files for a libstdc++
library, and as such only works in the shared scenario. When
linking statically to a libstdc++ library, the required printers
must be specified.

The default value of this option is 'none', if people feel 'auto'
is preferable then I'll change it.

Change-Id: Icdebd4f787387b99eaa60e0cf48058e9b8c484d4