ndk-build: Refine the handle of circular dependencies.
This changes the way circular dependencies are treated when computing
the topological sort in definitions-graph.mk.

Previously, only circular dependencies that cycled back to the
top-level module were handled properly, other ones were dropped.

For example, with an example like:

   A -> B
   B -> B

The module B was dropped from the result. This resulted in the
inability to use STLPORT_FORCE_REBUILD with stlport_static
(see bug entry below).

This implementation detects cycles at any level, and returns
the corresponding nodes, in their original listed order.

Since this modifies the result of the function when circular
dependencies are involved, had to modify one test case.

+ Fix typo.

https://code.google.com/p/android/issues/detail?id=56607

Change-Id: I20976d7d6d56ee0976dadc500d55dd34a8d7fa3b