libc++: Add testit_android script.
This patch adds a new shell script that can be used to
run the whole libc++ test suite on a connected Android device.

It is a copy of the libc++ 'testit' script, customized to build
with a standalone Android toolchain, and use ADB to run each test
on a customized device.

+ Change the names of the libc++ libraries to libc++_static.a
  and libc++_shared.so, both for the standalone and ndk-build
  builds.

+ Fix a bug where linking against the shared library built
  with build-cxx-stl.sh would fail to find nl_langinfo.

+ Fix a bug where build-cxx-stl.sh forgot to build the Android
  specific locale_android.cpp support file.

To run the test suite do the following:

  cd $NDK
  build/tools/build-cxx-stl.sh --stl=libc++
  build/tools/make-standalone-toolchain.sh --install-dir=/tmp/ndk-arm
  build/tools/make-standalone-toolchain.sh --install-dir=/tmp/ndk-x86 --arch=x86
  build/tools/make-standalone-toolchain.sh --install-dir=/tmp/ndk-mips --arch=mips
  cd sources/cxx-stl/llvm-libc++/libcxx/test
  PATH=$PATH:/tmp/ndk-arm/bin:/tmp/ndk-x86-bin:/tmp/ndk-mips/bin
  ./testit_android --abi=armeabi-v7a

For an Android/x86 device, replace last line with:
  ./testit_android --abi=x86

For an Android/MIPS one, replace last line with:
  ./testit_android --abi=mips

WARNING: Running the full test suite is *very* long, and several tests are
         failing at the moment.

Change-Id: I3f49e3a8fcea6f09f658092d864e3dcc65f33997