build-gcc.sh: Linkify binaries to reduce install size.
This is really useful to reduce the size of .zip archives
(about 7 MB / toolchain saved in the final .zip).

For Linux / Darwin, tar automatically detects the hard links
used by the toolchain install script and already copies a single
copy of each file in the archive (and will recreate the hard links
at install time), so this is less useful, but shouldn't hurt.

Change-Id: I9fd4309c6b7bd3c2cac7737c545b6e597d08ca8a