ndk-gdb.py: Fix ndk_bin_path determination
It wasn't correct on 64bit. This meant that the gnumake
found on those systems tended to be the system one.

Now, sys.executable is used and if that fails, some more
exhaustive checks of platform.platform() and sys.maxsize
are used (in-case the user uses their own Python - note
if this happens, arch differences between that Python and
the NDK could still cause the system make to be used, but
this is a small corner case that will usually work ok ..)