Error when gdb stub fail to build
In windows *-gdb.exe is a wrapper which setup up PATH and PYTHONHOME
to python prebuilt directory before calling into the real *-gdb-orig.exe.

For 32-bit windows, i686-w64-mingw32-gcc is used to compile gdb stub
because the default i586-mingw32msvc doesn't have functions such as
CreateJobObject used in the stub.  Problem is that i686-w64-mingw32-gcc
may not be installed in build system and the build script fails silently,
causing 32-bit *-gdb.exe fail to launch due to missing libpython2.7.dll.

This patch is to issue error when stub can't be generated for
whatever reason.

Change-Id: I6c939728aa7d90eafc60cb5637b598092557b584