Improve ndk-build output + fix double-quotes on Windows.
This patche slightly improves the ndk-build output for individual
build steps, i.e.:

  BEFORE:
     Compile <abi>  : <some details>

  AFTER:
     [abi] Compile  : <some details>

The reason for this is that there are some build steps which didn't
include the <abi>, which is ambiguous when doing a multi-ABI build with
lots of parallel jobs.

Note that for ARM, 'Compile arm' or 'Compile thumb' are still being
used to distinguish between 32-bit ARM and 16 or 16/32 Thumb compilation
modes. This only affects Compile and Compile++ build steps.

+ Fix the extra-quoting issue on Windows described
  by https://code.google.com/p/android/issues/detail?id=60649

Change-Id: Ib0bb7587f7113d488696e1c820d89adabad3070c