Changing docs for x86_64 ABI.
Note that some of the docs contain changes for compiler defaults which
are not effectively done yet. As soon as we finalize this I will submit
patches for corresponding compiler tuning.

In some of the docs I've put [TBD] instead of versions which can be
changed to some real numbers when time comes.

Change-Id: I7dc553fec267f2766244753dba982bc7018237e4
Signed-off-by: Pavel Chupin <pavel.v.chupin@intel.com>