Fix x86/mips breakage caused by kernel/uapi.
It seems that linux/stddef.h no longer defines NULL, and
crtbegin.o can not be compiled after the adoption of uapi.

This commit fix this issue by removing unnecessary isystem
and nostdinc option.  Now, the compilation command in
gen-platforms.sh should be similar to the build rules for
crtbegin.o in the bionic Android.mk.

Change-Id: Ic9c2588aac2d4f40054188f761cf95325a239094