[gabi++] nullptr can be converted to every pointer type.
The nullptr can be converted to every pointer type, so every catch
statements with pointer type should be able to catch the nullptr
exception.

This commit fixes following test cases for GCC 4.8/Clang 3.4:

* tests/device/test-gabi++/test_gabixx_shared_catch_const_pointer_nullptr
* tests/device/test-gabi++/test_gabixx_shared_catch_member_pointer_nullptr
* tests/device/test-gabi++/test_gabixx_shared_catch_pointer_nullptr
* tests/device/test-gabi++/test_gabixx_static_catch_const_pointer_nullptr
* tests/device/test-gabi++/test_gabixx_static_catch_member_pointer_nullptr
* tests/device/test-gabi++/test_gabixx_static_catch_pointer_nullptr

p.s. GCC 4.6 passed the test simply because the __cplusplus is not
implemented properly, and the test for C++11 (or C++0x) will always
pass.

Change-Id: I80bbaf7b57ddbc9104f659b5614f6c1277641aba