Try to remove libc++ dependencies on libcompiler-rt
Add --with-compiler-rt to testit_android and turn it off by default
Note that libc++ libraries are built with clang because it uses
clang-specific features.  Previous assumption is that libcompiler-rt
is also needed, which turns out not true.

Also note that traditionally clang in NDK use gcc's assembler,
linker, libgcc.a, etc, etc.  This change tries to extend the tradition
to libc++ land.

Change-Id: I86601636373a6fc6c7357fa733442f5be19b5593