docs: Update the list of return values from android_getCpuFamily
The existing docs were outdated.

Only mention the publicly supported ABIs for now, not all that
actually already are present in the source.

Adding a separate paragraph of text inbetween the list of return
values, to visually separate it from the next function in the
list of the two functions.

Change-Id: Icc1ddf112f99991d861a2f97df7568b00e0368b9