cpu-features: Support 32-bit ARM binaries running on an ARM64 kernel.
This addresses several issues related to 32-bit ARM binaries running
on an ARMv8 CPU.

The first one is that /proc/self/auxv should always be the preferred
way for a process to access the 32-bit ARM ELF HWCaps. Unfortunately,
this file is not always accessible on some recent Android platform
versions, even though this may be fixed in the future.

When the file is not accessible, parse /proc/cpuinfo as a fallback,
but deal with the fact that the 'Features' field only list features
that are optional for the device's CPU reference architecture revision.
I.e. when running on an ARMv8 CPU, it will not list the typical
ARMv7 extensions that become mandatory in ARMv8.

+ Refresh and clarify documentation to explain that the output of
  android_getCpuFamily() depends on the process' bitness, not the
  target device's CPU.

+ Check that android_getCpuFeatures() returns 0 as expected on
  anything except 32-bit ARM and 32-bit Intel.

BUG=13679666

Change-Id: I003808b0a44c0645a180232ab688f9ce76d10f25