Don't follow symlinks when generating shared library
If a shared library exists in one platform version, and gets
additional symbols in a later platform version, the gen-platforms.sh
script would incorrectly generate put a stub library with the new
symbols in the old platform. This happens because each platform is
initially populated with symlinks to files in earlier platforms.

When a new stub library is written, it should clobber the symlink, rather
than following the symlink and overwrite the stub library in the
earlier platform.

Change-Id: I0b6a7121ccbc2672311abf63931b329ee4e457ff