Exclude variables in libgcc.a from NDK *so
Shared libraries in NDK are generated from list of symbols
crated by running gen-system-symbols.sh on the real shared libraries.
The list may contain symbols merged from libgcc.a when the real
*so is linked, and should be removed.

Previously only functions in libgcc.a are excluded.  This CL
removes variables in libgcc.a as well.  Currently "__popcount_tab"
is only symbol affected

Change-Id: I3fb9fe712e958e9bf8879a7c115fe463916e160f