Remove sscanf and friends from android_support.a
*sscanf was added because 19 rand.dist classes in <random> use
'%a' to stream out/in bit-exact FP representation.  Unfortunately
*sscanf uses internal function __srefill not necessarily available in
the future libc.so.

It turns out that strtod is actually called to handle hexfloat
generated by printf("%a"), not *sscanf.  The strtod in android_support.a
is fixed to accept hexfloat.  So we are safe to remove it.

Support of "%a" in bionic's *scanf isn't available until android-21

Change-Id: Ia738e1040a5a95041eaade26c596663811326aef