Add --mclinker flag to build-llvm.sh
mclinker is only built with the default llvm version (which
is usually the most recent one of two llvm versions for each
NDK release).  Add --mclinker to force build of mclinker regardless
the llvm version.

Note that in theory mclinker doesn't have to be built with llvm
because in NDK mclinker is used just like a linker and accept only
archive of object files, not bitcode (as integrated linker in clang)
But because mclinker use llvm's header/libs a lot, it make sense to
build mclinker with clang/llvm when intermediate lib/obj are still
around (ie. to avoid rebuild llvm all over again just for mclinker)

Change-Id: Ic0291e06844deb2a0f28d290f564332b799ac56a