Call getauxval only when sdk.version >=20 in cpu-features
Function getauxval is available until API level 20. However
on ealier versions, that API might exist already but not behave
as expected. Checking of API level on target devices is to avoid
calling that API on those devices.

The dlopen/dlsym is kept instead of calling getauxval() directly.
This is because if that symbol does not existed, calling getauxval
directly will introduce relocation error in dynamic linker.

Change-Id: I3e1f10e39287eaa2ad6eb4d7cd059d96077f2ed4