Support using the .neon tag on files in arm64-v8a and x86_64 builds as well
Neon intrinsics work just fine on arm64-v8a, so source files with
such a tag should be usable on arm64-v8a right away (without
having to duplicate the source file definitions in Android.mk to
remove the .neon tag, even though it doesn't have any effect on
arm64-v8a).

The neon intrinsics header for x86 also works fine on x86_64, so
if the .neon suffix is allowed for x86, it should be allowed for
x86_64 just as well.

Change-Id: I2d8c4ce874fc134aa2abdc718d65d1bd84cc6d67