The default version of an application is now 16
If no APP_PLATFORM is specified in Application.mk, it is set now
to $(NDK_FIRST_PIE_PLATFORM_LEVEL) rather than 3 as it used to.

The reason behind this is to support by default new devices and not old,
so that a developer has to write less lines of config for the
application to run (also the error message about a PIE required is not
self-explaining at all). Furthermore most NDK tests do not have 'platform'
specified and thus failed when run on new android devices/emulator (with AOSP
latest from git master) because the executables were built as non-PIE.

Note, that after this patch most NDK tests will fail when run on an old
Android build (pre-4.1) apart from those with the platform explicitly
specified: they will not be launched at all.
If there is a need to really support tests on both old and new devices a
testing script and makefiles should be reworked in such a way that a
correct PIE/non-PIE version is built and run depending on the target
Android version.

Change-Id: Ibd611dff57a92f369fd9f154b6f6b15313f0e55f
Signed-off-by: Egor.Kochetov <egor.kochetov@intel.com>