Take LLVM from the platform prebuilts.
We won't be able to build the Windows LLVM until we get one in the
platform, but that doesn't build right now anyway.

As a side effect, stop building compiler-rt. We get this for free as
part of the prebuilt LLVM.

When building with LLVM, use the platform LLVM.

Note that having ANDROID_BUILD_TOP set is now a requirement. This can
be set manually, by `lunch`, or will be handled by `checkbuild.sh`,
but must be set.

Change-Id: Ieceace0e7b2091421eabb363f427f935ddd9436a