Fixed libdl_unittest for x86 in NDK self-tests
ndk test-android-support/libdl_unittest.c has been fixed so that it
expects a substring "/libc.so" but not a string "libc.so" as the library
name, since (as stated in the manpage) dli_fname is the pathname of the
library, not just the filename.

The test now passes, previously it was (incorrectly) failed.

See also: https://code.google.com/p/android/issues/detail?id=177387

Change-Id: I82cb01996913fee46ca946f7c77e6799bb449c67
Signed-off-by: Egor.Kochetov <egor.kochetov@intel.com>