Fixed NDK self-test unwind-struct on x86_64
An offset checking of one field of struct _Unwind_Exception was buggy
that caused the test to crash.
Specifically, there is a structure of the following form

struct _Unwind_Exception {
  uint64_t exception_class;
  _Unwind_Exception_Cleanup_Fn exception_cleanup;
  uint32_t private_1;
  uint32_t private_2;
}

in the file ndk/sources/cxx-stl/gabi++/include/unwind-itanium.h

The test checked that the private_1 field has offset 16 which is correct
for x86_64 architecture. The size of this field is 4 bytes, the
alignment is also 4 (according to x86_64 ABI specification).
Hence the next field will start right after the private_1 and will have
offset 20. The test expected it to be 24.
So it seems that the test is incoherent with reality and ABI
specifications. It is fixed by this patch.

BTW gcc4.8 and clang3.6 on the host x86_64 platform provide the same
results too, offset is 20.

See also: https://b.android.com/177423

Change-Id: Ieef84acea3ea4feabdc2d4dccf11229a0a934339
Signed-off-by: Egor Kochetov <egor.kochetov@intel.com>