Don't build target prebuilts for Mac or Windows.
We only need one copy of these, and they might as well come from Linux
since that's where we're working. At some point we should probably
transition to using the prebuilt toolchains to build the target
libraries, at which point the target libraries don't need to be
associated with the host build at all.

Change-Id: Ifb5d038e26e2af2ad30e6eef1fc0881a8d66e23f