Make checkbuild build 32-bit Windows binaries.
We still ship 32-bit Windows tools, so we can't force on --try-64
unconditionally. Disallow building Linux and Windows in the same
invocation because we don't have the control to build 32-bit and
64-bit at the same time.

As an added bonus, verify that the system we're passed is an actual
valid system. No more 20 minute typo penalties.

Change-Id: I53fbe16d2d50a26c19ed2aee0c2710dc22cc0749