Rewrite run-tests.sh in Python.
Giant shell scripts are difficult to maintain, and bash makes code
reuse clunky at best.

We already ship Python with the NDK, so this is actually more portable
than shipping a shell script.

The old test script will be left in place until we're sure the new
runner is sufficient.

Change-Id: I76325db6cd002a384ae6638b0575f1df2f6fc36a