WIP: Rewrite run-tests.sh in Python.
Giant shell scripts are difficult to maintain, and bash makes code
reuse clunky at best.

We already ship Python with the NDK, so this is actually more portable
than shipping a shell script.

The old test script will be left in place until we're sure the new
runner is sufficient.

Note that is is still a pretty rough edit. No need to dig too deep
yet. I'm mostly interested in knowing if I've missed any major use
cases or complicated the use of the test runner for anyone.

TODO: Integrate Egor's test statistics gathering.
This was a more or less pure translation of what we have today. It's
been written such that producing a report at the end shouldn't be
difficult, but I haven't done that yet.

TODO: Handle default ABI case.
The old test script would build and test all ABIs listed in APP_ABI in
a single invocation if an explicit ABI was not given. Currently this
will fall down over that case. I've written most of the code to handle
this, but I decided to factor that for-each up several levels because
it was going to make the device tests rather messy and I haven't
finished doing that yet.

TODO: Handle explicit test lists from command line.
The old test runner allowed specifying an exact list of tests to run
with --tests. That seems like a useful thing to keep around, but I
haven't ported it yet.

TODO: Split this into multiple modules.
There's quite a lot in this file, and there seem to be a few distinct
modules (adb, ndk-build, test types, the test runner).

Change-Id: I76325db6cd002a384ae6638b0575f1df2f6fc36a