Move ANDROID_BUILD_TOP not set warning.
Having this in prebuilt-common.sh meant that it had to be set for a
user to run make-standalone-toolchain.sh. The real bug here is that
make-standalone-toolchain.sh uses prebuilt-common.sh (for nothing but
the option parser), but prebuilt-common.sh and ndk-common.sh have
become far too monolithic to be split apart easily.

Moving it to rebuilt-all-prebuilt.sh doesn't cover much (running
build-host-gcc.sh won't warn now), but this is probably the best that
can be done quickly.

Change-Id: I5e7156ca7fb11cf926e52c6d5ea7864b92230da0