Don't try to unpack libgccunwind-libs on Darwin.
One of the misc fixes in
https://android-review.googlesource.com/#/c/68001/ was disabling
building libgccunwind for non-Linux builds. I don't know why this was
done, but it was, so let's unblock the build for now and investigate
enabling that later.

Note that we don't need to do this for Windows because the target libs
in the windows build are actually built as part of the Linux build.

Bug: http://b/22883654
Change-Id: I7d1d27c9e96938e70a8477b4b6cc50ac142a2857