Separate test discovery and execution.
This makes it easier to support running a subset of tests by name
because it lets us filter the test list at a high level rather than in
the guts of each test runner.

It also lets us perform some sanity checks on tests to catch
configuration errors before we start running (such as is now done with
the awk tests if they are missing a golden file).

Change-Id: I09f82bede5ac4d33f5b39d4ba99e654fed3174fb