use -fstack-check for GCC on x86 and x86_64
The -fstack-check switch prevents harmful stack overflows by enforcing
the guarantee that the guard page at the end of stacks will be hit. It
does this by injecting one byte writes for large uninitialized frames
and any usage of alloca. It rarely needs to do anything because these
cases are rare. Bionic includes a single guard page by default which is
all that's required for this to work.

Without this feature, even "correct" code can cause a stack overflow
overshooting the guard page. The most significant change is that an
unbounded alloca can only cause a denial of service since it will
reliability trigger a segfault. The stack protector feature can only
catch buffer overflows past the stack frame boundary, so it provides no
protection against these issues.

LLVM only supports this feature on Windows right now where it's a hard
requirement of the platform ABI. A more generic version has been
implemented but hasn't yet landed.

In GCC, this feature works well on x86 and x86_64 where there are robust
platform-specific implementations, but the generic fallback used on ARM
and some other architectures is broken:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65958

Bug: 13662332
Change-Id: Ieae859db44b43547783efe9430d7437621092558