Updated tests app_platform to run on Android 5.0+
For all the NDK device tests with the platform specified explicitly,
their
platform is now set to minimal_pie_platform, that is 16. This way, the
tests
will be able to run on all new devices and will be skipped on the old
ones.
Since android <4.1 and android >5.0 use incompatible executables (-fPIE)
it is a required change to run tests on new devices without reworking
all tests' makefiles and the test script -- in which case both new and
old devices would be supported.

See also: http://b.android.com/177525

Change-Id: If3369bcd14905dde1cc088c2bba4818bc086aa0a
Signed-off-by: Egor.Kochetov <egor.kochetov@intel.com>
Signed-off-by: Egor Kochetov <egor.kochetov@intel.com>