Improve adb shell status search.
Apparently on android-16 the linker (or perhaps adbd) would emit a
failure in the form of "ERROR MESSAGE\r\n\0", so we will end up using
"\014" (for a return status of 14), and then we won't be able to parse
the int because of the null.

This moves us closer to the implementation in the true ADB python
interface. We should move to just using that, but first we'll need to
actually package that up so it can be shipped in the NDK.

Change-Id: If16610d802978fb6715f08b40fbf254c7cfe3a22