Fix detection of thumb STL libraries.
I think the CWD must have changed when I moved us to using
build_support.py, which means we no longer find the thumb libraries
and don't try to package them.

Note that we don't need to correct the paths that we actually add for
packaging because pack_archive will cd into the right directory for
that operation anyway.

Change-Id: Ia8ee9bc056d8570829f77b0c19d050a71cb564ea