Add test_config.json for better test disabling.
Each test directory can contain a "test_config.json" file that
describes the configurations a test is not expected to pass for.
Previously this information could be captured in one of two places:
the Application.mk file, or a BROKEN_BUILD/BROKEN_RUN file.

Application.mk was used to state that a test was only to be run for a
specific platform version, specific toolchain, or a set of ABIs.
Unfortunately Application.mk could only specify a single toolchain or
platform, not a set.

BROKEN_BUILD/BROKEN_RUN files were too general. An empty file meant
the test should always be skipped regardless of configuration. Any
change that would put a test in that situation should be reverted
immediately. These also didn't make it clear if the test was actually
broken (and thus should be fixed) or just not applicable.

A test_config.json file is more descriptive. It can mark a test as
"broken" or "unsupported", and can do so with a list of "abis",
"platforms", or "toolchains". It is assumed that most tests are valid
for any configuration. Such tests require no test_config.json file.

Change-Id: Ibffcef610a32c49efef17eba1154f6a9f4dcf565