Stop packaging the out of date samples.
The samples in development/ndk are not maintained, and some of them
don't work (gles3jni, for example).

In a modular NDK world we'll be shipping these directly from github
anyway.

Note that this also removes the samples from the tests. We should
probably find a way to work these back in using the latest samples
from github, but for now they aren't actually helpful tests.

Change-Id: I073df17fd2256b8f2105adaccf0a7e877716a60b